+++ tests/testthat/test-plot_genoprob.R
@@ -2,7 +2,7 @@ context("plot_genoprob")
 
 test_that("plot_genoprob works", {
 
-    skip_if(isnt_karl(), "plot tests only done locally")
+    skip_if(isnt_karl(), "plot tests only run locally")
 
     iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2"))
     iron <- iron[c("116", "232"),2]