+++ tests/testthat/test-plot_peaks.R
@@ -2,7 +2,7 @@ context("plot_peaks")
 
 test_that("plot_peaks works", {
 
-    skip_if(isnt_karl(), "plot tests only done locally")
+    skip_if(isnt_karl(), "plot tests only run locally")
 
     iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2"))
     map <- insert_pseudomarkers(iron$gmap, step=1)