+++ tests/testthat/test-scan1perm.R
@@ -132,7 +132,7 @@ test_that("scan1 permutations work (regression test)", {
 
 test_that("scan1 permutations work with single kinship matrix (regression test)", {
 
-    skip_if(isnt_karl(), "This test only run locally")
+    skip_if(isnt_karl(), "this test only run locally")
 
     seed <- 3025685
     RNGkind("L'Ecuyer-CMRG")
@@ -249,7 +249,7 @@ test_that("scan1 permutations work with single kinship matrix (regression test)"
 
 test_that("scan1 permutations work with LOCO kinship matrix (regression test)", {
 
-    skip_if(isnt_karl(), "This test only run locally")
+    skip_if(isnt_karl(), "this test only run locally")
 
     seed <- 3025685
     RNGkind("L'Ecuyer-CMRG")